+++ BucketList/BucketList/ViewControllers/Posts/PostCollectionViewCell.swift
@@ -1,0 +1,18 @@
+//
+//  PostCollectionViewCell.swift
+//  BucketList
+//
+//  Created by Lambda_School_Loaner_188 on 10/22/19.
+//  Copyright © 2019 Stephanie Bowles. All rights reserved.
+//
+
+import UIKit
+
+class PostCollectionViewCell: UICollectionViewCell {
+    
+    //Outlets
+    @IBOutlet weak var imageCategory: UIImageView!
+    @IBOutlet weak var titleCategoryLabel: UILabel!
+    
+    
+}
